Fixes null clerkUserMap
